Dodalem potwierdzenie usuniecia pojazdu i kilka bugfixów :)

+++ src/views/Search.vue
@@ -11,19 +11,41 @@
           <ion-title size="large">Database</ion-title>
         </ion-toolbar>
       </ion-header>   
+      <ion-button @click="onClickedReloadPage">
+       <ion-icon size="large" src="../assets/icons/refresh-circle-outline.svg"/>
+      </ion-button>
         <div id="container">
           <ion-list v-for="car in carBase" :key="car.id">
             <ion-item-sliding>
               <ion-item>
                 <ion-label class="ion-text-wrap">
-                  Brand: {{car.marka}} Model: {{car.model}} Price: {{car.cena}}<br/>
-                  Year: {{car.rocznik}} Mileage: {{car.przebieg}} Power: {{car.moc}}<br/>
-                  Doors: {{car.drzwi}} Color: {{car.kolor}}
+                  Brand: {{car.marka}} 
+                  <br/>
+                  Model: {{car.model}} 
+                  <br/>
+                  Price: {{car.cena}}
+                  <br/>
+                  Year: {{car.rocznik}} 
+                  <br/>
+                  Mileage: {{car.przebieg}} 
+                  <br/>
+                  Power: {{car.moc}}
+                  <br/>
+                  Doors: {{car.drzwi}} 
+                  <br/>
+                  Color: {{car.kolor}}
                 </ion-label>
               </ion-item>
               <ion-item-options side="end">
-                <ion-item-option @click="onDeleteClicked(car.id)" color="danger">Delete</ion-item-option>
+                <ion-item-option @click="onDeleteHandler(car.id)" color="danger">Delete</ion-item-option>
               </ion-item-options>
+              <ion-alert
+                :is-open="isOpenRef"
+                header="Delete"
+                message="Confirm this alert to delete selected entry"
+                :buttons="buttons"
+                @didDismiss="setOpen(false)"
+              />
             </ion-item-sliding>
           </ion-list>
         </div>
@@ -32,17 +54,36 @@
 </template>
 
 <script>
-import { IonPage, IonHeader, IonToolbar, IonTitle, IonContent, IonItem, IonList, IonLabel  } from '@ionic/vue';
+import { IonPage, IonHeader, IonToolbar, IonTitle, IonContent, IonItem, IonList, IonLabel, IonItemSliding, IonItemOption, IonItemOptions, IonIcon, IonButton, IonAlert  } from '@ionic/vue';
 
 export default  {
   name: 'Search',
-  components: { IonHeader, IonToolbar, IonTitle, IonContent, IonPage, IonItem, IonList, IonLabel },
+  components: { IonHeader, IonToolbar, IonTitle, IonContent, IonPage, IonItem, IonList, IonLabel, IonItemSliding, IonItemOption, IonItemOptions, IonIcon, IonButton, IonAlert },
   data() {
     return {
+      isOpenRef: false,
+      setOpen: (state =  Boolean.prototype) => this.isOpenRef = state,
+      idToDelete: null,
+      buttons: [
+            {
+              text: 'Cancel',
+              role: 'cancel'
+            },
+            {
+              text: 'Confirm',
+              handler: () => {
+                this.onDeleteClicked(this.idToDelete)
+              },
+            },
+          ],
       carBase: []
     }
   },
   methods: {
+    onDeleteHandler(id) {
+      this.idToDelete = id;
+      this.setOpen();
+    },
     search() {
       fetch("http://localhost:3000/baza").then((res) => {
         res
@@ -52,16 +93,21 @@ export default  {
       });
     },
     onDeleteClicked(id) {
-      fetch("http://localhost:3000/baza/"+id, {
-        method: 'DELETE'
-      })
-      .then(response => response.json())
-      .then(data => {
-        console.log('Success:', data);
-      })
-      .catch((error) => {
-        console.error('Error:', error);
-      });
+      if(id != null){
+        fetch("http://localhost:3000/baza/"+id, {
+          method: 'DELETE'
+        })
+        .then(response => response.json())
+        .then(data => {
+          console.log('Success:', data);
+        })
+        .catch((error) => {
+          console.error('Error:', error);
+        });
+        window.location.reload();
+      }
+    },
+    onClickedReloadPage() {
       window.location.reload();
     }
   },
